requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
crispy-bootstrap5==2024.2
cryptography==43.0.1
distro==1.9.0
Django==5.1.1
django-allauth==64.2.1
django-bootstrap-modal-forms==3.0.4
django-crispy-forms==2.3
django-rest-framework==0.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
importlib_resources==6.4.5
inflection==0.5.1
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
oauthlib==3.2.2
openai==1.45.0
packaging==24.1
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
setuptools==74.1.2
sniffio==1.3.1
sqlparse==0.5.1
swagger-spec-validator==3.0.4
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
